func_801C3708

+++ src/st/np3/np3.h
@@ -3,8 +3,8 @@
 void CreateEntityFromCurrentEntity(u16, Entity*);
 void ReplaceBreakableWithItemDrop(Entity*);
 s32 func_801BCB5C(u16* arg0);
-s16 func_801BD308(u16* arg0, s16 arg1);
-s16 func_801BC7D4(void);
+s16 func_801BD308(u16* arg0, s32 arg1);
+s32 func_801BC7D4(void);
 int func_801CD658();
 void EntityPriceDrop(Entity* entity);
 void EntityInventoryDrop(Entity* entity);
